Renamed files to make sure they are goose specific
