+++ goose_utils.py
