+++ goose_version.py
